Move common routines from hw_X11.c and hw_xft.c into hw_x/common.c
