+++ server/hw/hw_x/flavor_protos.h
@@ -1,0 +1,3 @@
+/* Function prototypes for functions that each X11-based hw_xyz.c must implement. */
+
+static char * X11_AutodetectFont(udat fontwidth, udat fontheight);
